Rename the module name to swift.logic
Close: #14
